+++ Sources/WalletConnectPairing/PairingClientFactory.swift
@@ -21,7 +21,9 @@ public struct PairingClientFactory {
 
         let walletPaS = WalletPairService(networkingInteractor: networkingInteractor, kms: kms, pairingStorage: pairingStore)
 
-        return PairingClient(appPairService: appPairService, networkingInteractor: networkingInteractor, logger: logger, walletPairService: walletPaS, socketConnectionStatusPublisher: relayClient.socketConnectionStatusPublisher)
+        let pairingRequestsSubscriber = PairingRequestsSubscriber(networkingInteractor: networkingInteractor, logger: logger, kms: kms)
+
+        return PairingClient(appPairService: appPairService, networkingInteractor: networkingInteractor, logger: logger, walletPairService: walletPaS, pairingRequestsSubscriber: pairingRequestsSubscriber, socketConnectionStatusPublisher: relayClient.socketConnectionStatusPublisher)
     }
 }
